+++ mall-mbg/src/main/java/com/macro/mall/mapper/PmsProductCategoryMapper.java
@@ -12,7 +12,7 @@ public interface PmsProductCategoryMapper {
 
     int deleteByExample(PmsProductCategoryExample example);
 
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(PmsProductCategory record);
